discord-tools/requirements.txt
discord.py>=2.0.0
google-api-python-client>=2.0.0
google-auth>=2.0.0
google-auth-oauthlib>=0.4.0
python-dotenv>=0.19.0
matplotlib